requirements.txt
pyzmq
RPi.GPIO